pyproject.toml
[project]
name = "nonebot-plugin-pjsk"
version = "0.2.8"
description = "Project Sekai Sticker Creator for NoneBot2."
authors = [
    { name = "Agnes_Digital", email = "Z735803792@163.com" },
    { name = "student_2333", email = "lgc2333@126.com" },
]
dependencies = [
    "nonebot2>=2.0.0",
]
requires-python = ">=3.9,<4.0"
keywords = ["pjsk", "nonebot2", "plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/Agnes4m/nonebot_plugin_pjsk"

[tool.pdm.build]
includes = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pyright]
pythonVersion = "3.8"

[tool.black]
line-length = 89
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
# skip-string-normalization = true

[tool.ruff.isort]
extra-standard-library = ["typing_extensions"]

[tool.ruff]
ignore = [
    "B008",
    "B905",
    "E501",
    "FBT002",
    "PGH003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF100",
    "SIM117",
    "TRY002",
    "TRY003",
]
select = [
    "A",
    "ARG",
    "ASYNC",
    "B",
    "C4",
    "COM",
    "E",
    "F",
    "FBT",
    "FLY",
    "I",
    "ISC",
    "N",
    "PIE",
    "PGH",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "TRY",
]
